+++ src/ReplicatedStorage/Animations/Package/Util/AnimationsClass.lua
@@ -9,20 +9,24 @@ local Signal = require(script.Parent.Signal)
 local Await = require(script.Parent.Await)
 local ChildFromPath = require(script.Parent.ChildFromPath)
 local CustomAssert = require(script.Parent.CustomAssert)
+local AnimatedObjectsFolder = script.Parent.Parent.Parent.Deps.AnimatedObjects
+
+local ANIMATED_OBJECT_MOTOR6D_NAME = "AnimatedObjectMotor6D"
+local ANIM_ID_STR = "rbxassetid://%i"
 
 local function getRig(player_or_rig)
 	local playerCharacter 
-	
+
 	if player_or_rig:IsA("Player") then
 		task.delay(5, function()
 			if not playerCharacter then -- Can happen if `Players.CharacterAutoLoads` is disabled and this function gets called before it has a chance to load in
 				warn("Infinite yield possible on 'player_or_rig.CharacterAdded:Wait()'")
 			end
 		end)
-		
+
 		playerCharacter = player_or_rig.Character or player_or_rig.CharacterAdded:Wait()
 	end
-	
+
 	return playerCharacter or player_or_rig
 end
 
@@ -41,9 +45,21 @@ end
 local function createAnimationInstance(animationName, animationId)
 	local animation = Instance.new("Animation")
 	animation.Name = animationName
-	animation.AnimationId = "rbxassetid://" .. animationId
+	animation.AnimationId = ANIM_ID_STR:format(animationId)
 
 	return animation
+end
+
+local function setToolRightGripWeldEnabled(rig, tool, enabled)
+	if not tool:IsA("Tool") then
+		return
+	end
+	
+	local rightGripWeld = (rig:FindFirstChild("Right Arm") or rig:FindFirstChild("RightHand")):FindFirstChild("RightGrip")
+	
+	if rightGripWeld then
+		rightGripWeld.Enabled = enabled
+	end	
 end
 
 local AnimationsClass = {}
@@ -53,18 +69,56 @@ AnimationsClass.__index = AnimationsClass
 -- INITIALIZATION --
 --------------------
 function AnimationsClass:_animationIdsToInstances()
+	local function replaceAnimationIdWithAnimationInstance(idTable, animationIndex, animationId)
+		self.AnimationInstancesCache[animationId] = self.AnimationInstancesCache[animationId] or createAnimationInstance(tostring(animationIndex), animationId)
+
+		idTable[animationIndex] = self.AnimationInstancesCache[animationId]
+	end
+
+	local function mapAnimationInstanceToAnimatedObjectInfo(animationInstance, animatedObjectInfo)
+		local oldAnimatedObjectPath = self.AnimationInstanceToAnimatedObjectInfo[animationInstance]
+
+		if oldAnimatedObjectPath ~= nil then
+			CustomAssert(oldAnimatedObjectPath == animatedObjectInfo, "Conflicting `animatedObjectInfo` for `animationId` [", animationInstance.AnimationId, "]")
+		else
+			self.AnimationInstanceToAnimatedObjectInfo[animationInstance] = animatedObjectInfo
+		end
+	end
+
+	local function mapAnimationInstancesToAnimatedObjectInfo(animationInstancesTable, animatedObjectInfo)
+		-- `idTable` becomes `animationInstancesTable` after calling `replaceAnimationIdWithAnimationInstance` on key-value pairs
+		for _, animationInstance in pairs(animationInstancesTable) do
+			if type(animationInstance) == "table" then
+				local nextIdTable = animationInstance
+
+				mapAnimationInstancesToAnimatedObjectInfo(nextIdTable, animatedObjectInfo)
+			else
+				mapAnimationInstanceToAnimatedObjectInfo(animationInstance, animatedObjectInfo)
+			end
+		end
+	end
+
 	local function loadAnimations(idTable)
-		for animationName, animationId in pairs(idTable) do
+		for animationIndex, animationId in pairs(idTable) do
 			if type(animationId) == "table" then
 				local nextIdTable = animationId
-				
-				loadAnimations(nextIdTable)
-			else
-				if type(animationId) ~= "userdata" then -- The `animationId` was already turned into an animation instance (happens when multiple references to the same animation id table occur)
-					self.AnimationInstancesCache[animationId] = self.AnimationInstancesCache[animationId] or createAnimationInstance(tostring(animationName), animationId)
+				local animatedObjectInfo = nextIdTable._animatedObjectInfo
 
-					idTable[animationName] = self.AnimationInstancesCache[animationId]
+				if animatedObjectInfo then
+					nextIdTable._animatedObjectInfo = nil
+
+					if nextIdTable._singleAnimationId then -- If there is only 1 animation id in the table, just replace the table itself with an animation instance (also don't need to worry about setting to nil because the whole table gets overloaded with the `_singleAnimationId`)
+						replaceAnimationIdWithAnimationInstance(idTable, animationIndex, nextIdTable._singleAnimationId)
+						mapAnimationInstanceToAnimatedObjectInfo(idTable[animationIndex], animatedObjectInfo)
+					else
+						loadAnimations(nextIdTable)
+						mapAnimationInstancesToAnimatedObjectInfo(nextIdTable, animatedObjectInfo)
+					end
+				else
+					loadAnimations(nextIdTable)
 				end
+			elseif type(animationId) ~= "userdata" then -- The `animationId` was already turned into an animation instance (happens when multiple references to the same animation id table occur)
+				replaceAnimationIdWithAnimationInstance(idTable, animationIndex, animationId)
 			end
 		end
 	end
@@ -82,12 +136,12 @@ function AnimationsClass:_createInitializedAssertionFn()
 	else
 		instanceName = "AnimationsClient"
 	end
-	
+
 	self._initializedAssertion = function()
 		assert(self._initialized, "Call " .. instanceName .. ":Init() before calling any methods")
 	end
 end
-	
+
 -----------------
 -- CONSTRUCTOR --
 -----------------
@@ -97,8 +151,10 @@ function AnimationsClass.new()
 	self.FinishedLoadingRigSignal = Signal.new()
 	self.LoadedTracks = {}
 	self.Aliases = {}
+	self.AnimationInstanceToAnimatedObjectInfo = {}
+	self.TrackToAnimatedObjectInfo = {}
 	self.AnimationInstancesCache = {}
-	self.IsRigLoaded = {}	
+	self.IsRigLoaded = {}
 
 	self:_animationIdsToInstances()
 	self:_createInitializedAssertionFn()
@@ -122,23 +178,9 @@ function AnimationsClass:_getTrack(player_or_rig, path, isAlias)
 		parent = self.LoadedTracks[rig]
 	end
 
-	--local firstKey = path
+	local track_or_id_table = ChildFromPath(parent, path)
 
-	--if typeof(path) == "table" then
-	--	if #path > 1 then
-	--		firstKey = path[1]
-	--	end
-	--else
-	--	if path:match("%.") then
-	--		firstKey = path:match("^[^%.]+")
-	--	end
-	--end
-
-	--CustomAssert(parent, "no track or table found under path [", path, "] for player or rig [", player_or_rig:GetFullName(), "]")
-	
-	local track_or_table = ChildFromPath(parent, path)
-
-	return track_or_table
+	return track_or_id_table
 end
 
 function AnimationsClass:_playStopTrack(play_stop, player_or_rig, path, isAlias, fadeTime, weight, speed)
@@ -167,11 +209,11 @@ function AnimationsClass:_setTrackAlias(player_or_rig, alias, path)
 	local track_or_table = self:_getTrack(player_or_rig, modifiedPath) or self:_getTrack(player_or_rig, path)	
 
 	CustomAssert(track_or_table, "No track or table found under path [", path, "] with alias [", alias, "] for [", player_or_rig:GetFullName(), "]")
-	
+
 	if not self.Aliases[player_or_rig] then
 		self.Aliases[player_or_rig] = {}
 	end
-	
+
 	self.Aliases[player_or_rig][alias] = track_or_table
 end
 
@@ -193,12 +235,166 @@ function AnimationsClass:_getAnimator(player_or_rig, rig)
 	return animator
 end
 
+function AnimationsClass:_attachDetachAnimatedObject(attach_detach, player_or_rig, animatedObjectSourcePath_or_animationTrack_or_animatedObject)
+	local rig = getRig(player_or_rig)
+
+	local animatedObjectSourcePath = nil
+
+	if typeof(animatedObjectSourcePath_or_animationTrack_or_animatedObject) == "Instance" then
+		if animatedObjectSourcePath_or_animationTrack_or_animatedObject:IsA("AnimationTrack") then
+			local animationTrack = animatedObjectSourcePath_or_animationTrack_or_animatedObject
+
+			animatedObjectSourcePath = self.TrackToAnimatedObjectInfo[rig][animationTrack.Animation.AnimationId].AnimatedObjectPath
+		end
+	else
+		animatedObjectSourcePath = animatedObjectSourcePath_or_animationTrack_or_animatedObject
+	end
+
+	local animatedObjectSource 
+		
+	if animatedObjectSourcePath then
+		animatedObjectSource = ChildFromPath(AnimatedObjectsFolder, animatedObjectSourcePath)
+	else
+		animatedObjectSource = animatedObjectSourcePath_or_animationTrack_or_animatedObject
+	end
+
+	CustomAssert(animatedObjectSource, "No animated objects found under path [", animatedObjectSourcePath, "]")
+
+	local animatedObjectSourceIsMotor6D = animatedObjectSource:IsA("Motor6D")
+
+	local animatedObjectMotor6DSource = animatedObjectSourceIsMotor6D and animatedObjectSource or animatedObjectSource:FindFirstChild(ANIMATED_OBJECT_MOTOR6D_NAME, true)
+
+	if not animatedObjectSourceIsMotor6D then
+		CustomAssert(animatedObjectMotor6DSource, "No AnimatedObjectMotor6D (\"" .. ANIMATED_OBJECT_MOTOR6D_NAME .. "\") found for animated object [", animatedObjectSource:GetFullName(), "]")
+	end
+
+	local part0NameStringValue = animatedObjectMotor6DSource:FindFirstChild("Part0Name")
+	local part0Name = part0NameStringValue and part0NameStringValue.Value
+
+	CustomAssert(part0Name, "No AnimatedObjectMotor6D.Part0Name found for [", animatedObjectMotor6DSource:GetFullName(), "]")
+
+	local part1NameStringValue = animatedObjectMotor6DSource:FindFirstChild("Part1Name")
+	local part1Name = part1NameStringValue and part1NameStringValue.Value
+
+	CustomAssert(part1Name, "No AnimatedObjectMotor6D.Part1Name found for [", animatedObjectMotor6DSource:GetFullName(), "]")
+
+	local attachedAnimatedObject = rig:FindFirstChild(animatedObjectSource.Name)
+	local attachedAnimatedObjectMotor6D = attachedAnimatedObject and attachedAnimatedObject:FindFirstChild(ANIMATED_OBJECT_MOTOR6D_NAME)
+
+	local function attach(animatedObjectSource, parent)
+		local animatedObject
+
+		if animatedObjectSourceIsMotor6D then
+			animatedObject = attachedAnimatedObject
+		else
+			animatedObject = animatedObjectSource:Clone()
+			animatedObject.Parent = parent
+		end
+
+		setToolRightGripWeldEnabled(rig, animatedObject, false)
+
+		local animatedObjectMotor6D
+
+		if animatedObjectSourceIsMotor6D then
+			animatedObjectMotor6D = animatedObjectSource:Clone()
+			animatedObjectMotor6D.Name = ANIMATED_OBJECT_MOTOR6D_NAME
+			animatedObjectMotor6D.Parent = animatedObject
+		else
+			animatedObjectMotor6D = animatedObject:FindFirstChild(ANIMATED_OBJECT_MOTOR6D_NAME, true)
+		end
+
+		-- Setting the motor6D's Part0 to the correct body part so it will work with the animation
+		animatedObjectMotor6D.Part0 = rig:FindFirstChild(part0Name, true)
+
+		-- Setting the motor6D's Part1 to the correct tool part so it will work with the animation
+		animatedObjectMotor6D.Part1 = animatedObject.Name == part1Name and animatedObject or animatedObject:FindFirstChild(part1Name, true)	
+	end
+
+	if attach_detach == "detach" then
+		if attachedAnimatedObject then
+			setToolRightGripWeldEnabled(rig, attachedAnimatedObject, true)
+
+			if animatedObjectSourceIsMotor6D then
+				if attachedAnimatedObjectMotor6D then
+					if self.AnimatedObjectsDebugMode then
+						print(`Detaching AnimatedObjectMotor6D [ {attachedAnimatedObjectMotor6D:GetFullName()} ]`)
+					end
+
+					attachedAnimatedObjectMotor6D:Destroy()
+				end
+			else
+				if self.AnimatedObjectsDebugMode then
+					print(`Detaching animated object [ {attachedAnimatedObject:GetFullName()} ]`)
+				end
+
+				attachedAnimatedObject:Destroy()
+			end
+		end
+	elseif attach_detach == "attach" then
+		if animatedObjectSourceIsMotor6D then
+			if attachedAnimatedObjectMotor6D then
+				if self.AnimatedObjectsDebugMode then
+					print(`Unable to attach AnimatedObjectMotor6D [ {animatedObjectMotor6DSource.Name} ] because [ {attachedAnimatedObjectMotor6D:GetFullName()} ] is already attached`)
+				end
+
+				return
+			end
+
+			if not attachedAnimatedObject then
+				if self.AnimatedObjectsDebugMode then
+					print(`Unable to attach AnimatedObjectMotor6D [ {animatedObjectMotor6DSource.Name} ] because [ {animatedObjectMotor6DSource.Name} ] is not an attached animated object (child of the rig)`)
+				end
+
+				return
+			end
+		elseif attachedAnimatedObject then
+			if self.AnimatedObjectsDebugMode then
+				print(`Unable to attach animated object [ {animatedObjectSource.Name} ] because [ {attachedAnimatedObject:GetFullName()} ] is already attached`)
+			end
+
+			return
+		end
+
+		if self.AnimatedObjectsDebugMode then
+			if animatedObjectSourceIsMotor6D then
+				print(`Attaching AnimatedObjectMotor6D [ {animatedObjectMotor6DSource.Name} ]`)
+			else
+				print(`Attaching animated object [ {animatedObjectSource.Name} ]`)
+			end
+		end
+
+		if animatedObjectSource:IsA("Folder") then
+			local animatedObjectFolder = Instance.new("Folder")
+			animatedObjectFolder.Name = animatedObjectSource.Name
+			animatedObjectFolder.Parent = rig
+
+			for _, animatedObjectSourceChild in ipairs(animatedObjectSource:GetChildren()) do
+				attach(animatedObjectSourceChild, animatedObjectFolder)
+			end
+		else
+			attach(animatedObjectSource, rig)
+		end
+	end
+end
+
 -------------
 -- PUBLIC --
 -------------
+function AnimationsClass:AttachAnimatedObject(player_or_rig: Player | Model, animatedObjectSourcePath_or_animationTrack_or_animatedObject: ({any} | string) | AnimationTrack | any)
+	self._initializedAssertion()
+		
+	self:_attachDetachAnimatedObject("attach", player_or_rig, animatedObjectSourcePath_or_animationTrack_or_animatedObject)
+end
+
+function AnimationsClass:DetachAnimatedObject(player_or_rig: Player | Model, animatedObjectSourcePath_or_animationTrack_or_animatedObject: ({any} | string) | AnimationTrack | any)
+	self._initializedAssertion()
+
+	self:_attachDetachAnimatedObject("detach", player_or_rig, animatedObjectSourcePath_or_animationTrack_or_animatedObject)
+end
+
 function AnimationsClass:AwaitLoaded(player_or_rig: Player | Model)
 	self._initializedAssertion()
-	
+
 	local rig = getRig(player_or_rig)
 
 	if not self.IsRigLoaded[rig] then
@@ -208,13 +404,13 @@ end
 
 function AnimationsClass:AreTracksLoaded(player_or_rig: Player | Model): boolean
 	self._initializedAssertion()
-	
+
 	return not not self.LoadedTracks[getRig(player_or_rig)]
 end
 
 function AnimationsClass:LoadTracks(player_or_rig: Player | Model, rigType: string)
 	self._initializedAssertion()
-	
+
 	CustomAssert(AnimationIds[rigType], "No animations found for [", player_or_rig:GetFullName(), "] under rig type [", rigType, "]")
 
 	local rig = getRig(player_or_rig)
@@ -222,9 +418,9 @@ function AnimationsClass:LoadTracks(player_or_rig: Player | Model, rigType: stri
 	rig:SetAttribute("AnimationsRigType", rigType)
 
 	CustomAssert(not self.IsRigLoaded[rig], "Animation tracks already loaded for [", rig:GetFullName(), "] !")
-	
-	local animator = self:_getAnimator(player_or_rig, rig)
-	
+
+	local animator: Animator = self:_getAnimator(player_or_rig, rig)
+
 	local tracks = self.LoadedTracks[rig]
 
 	if not tracks then
@@ -240,78 +436,127 @@ function AnimationsClass:LoadTracks(player_or_rig: Player | Model, rigType: stri
 	end
 
 	local function loadTracks(animationsTable, currentSet)
-		for animationName, animation in pairs(animationsTable) do
-			if type(animation) == "table" then
+		for animationName, animationInstance in pairs(animationsTable) do
+			if type(animationInstance) == "table" then
 				currentSet[animationName] = {}
 
-				loadTracks(animation, currentSet[animationName])
+				loadTracks(animationInstance, currentSet[animationName])
 			else
-				ContentProvider:PreloadAsync({animation})
-				currentSet[animationName] = animator:LoadAnimation(animation)
+				ContentProvider:PreloadAsync({animationInstance})
+
+				local track = animator:LoadAnimation(animationInstance)
+
+				currentSet[animationName] = track
+
+				local animatedObjectInfo = self.AnimationInstanceToAnimatedObjectInfo[animationInstance]
+
+				if animatedObjectInfo then
+					self.TrackToAnimatedObjectInfo[rig][track.Animation.AnimationId] = animatedObjectInfo
+				end
 			end
 		end
 	end
 
+	self.TrackToAnimatedObjectInfo[rig] = {}
+
 	loadTracks(AnimationIds[rigType], self.LoadedTracks[rig])
 
+	animator.AnimationPlayed:Connect(function(track) -- When a track gets played attach an animated object if required
+		if self.AnimatedObjectsDebugMode then
+			print(`Animation track [ {track.Animation.AnimationId:match("%d.*")} ] played on rig [ {rig.Name} ]`)
+		end
+		
+		local animatedObjectInfo = self.TrackToAnimatedObjectInfo[rig][track.Animation.AnimationId]
+
+		if animatedObjectInfo then
+			local autoAttachDetachSettings = animatedObjectInfo.AutoAttachDetachSettings
+			
+			if autoAttachDetachSettings then
+				local runContext = autoAttachDetachSettings.RunContext
+				
+				CustomAssert(runContext == "Client" or runContext == "Server", "Invalid AutoAttachDetachSettings.RunContext [", runContext,"] for animation track [", track.Animation.AnimationId:match("%d.*"), "]")
+				
+				local isCorrectRunContext = RunService["Is" .. runContext](RunService)
+				
+				if isCorrectRunContext then
+					if autoAttachDetachSettings.AutoAttach then
+						self:_attachDetachAnimatedObject("attach", player_or_rig, animatedObjectInfo.AnimatedObjectPath)
+					end
+
+					if autoAttachDetachSettings.AutoDetach then
+						track.Stopped:Once(function()
+							if self.AnimatedObjectsDebugMode then
+								print(`Animation track [ {track.Animation.AnimationId:match("%d.*")} ] stopped on rig [ {rig.Name} ]`)
+							end
+
+							self:_attachDetachAnimatedObject("detach", player_or_rig, animatedObjectInfo.AnimatedObjectPath)
+						end)
+					end
+				end
+			end
+		end
+	end)
+	
 	self.IsRigLoaded[rig] = true
 
 	rig.Destroying:Connect(function() -- When the rig gets destroyed the animation tracks are no longer loaded
 		self.IsRigLoaded[rig] = false
+		self.LoadedTracks[rig] = nil -- Release memory
+		self.TrackToAnimatedObjectInfo[rig] = nil -- Release memory
 	end)
 
 	if self.TimeToLoadPrints then
 		print("Finished pre-loading animations for [", rig, "] -", os.clock()-s, "seconds taken")
 	end
-	
+
 	self.FinishedLoadingRigSignal:Fire(rig)
 end
 
-function AnimationsClass:GetTrack(player_or_rig: Player | Model, path: any): AnimationTrack?
+function AnimationsClass:GetTrack(player_or_rig: Player | Model, path: {any} | string): AnimationTrack?
 	self._initializedAssertion()
-	
+
 	return self:_getTrack(player_or_rig, path)
 end
 
-function AnimationsClass:PlayTrack(player_or_rig: Player | Model, path: any, fadeTime: number?, weight: number?, speed: number?): AnimationTrack
+function AnimationsClass:PlayTrack(player_or_rig: Player | Model, path: {any} | string, fadeTime: number?, weight: number?, speed: number?): AnimationTrack
 	self._initializedAssertion()
-	
+
 	return self:_playStopTrack("Play", player_or_rig, path, false, fadeTime, weight, speed)
 end
 
-function AnimationsClass:StopTrack(player_or_rig: Player | Model, path: any, fadeTime: number?): AnimationTrack
+function AnimationsClass:StopTrack(player_or_rig: Player | Model, path: {any} | string, fadeTime: number?): AnimationTrack
 	self._initializedAssertion()
-	
+
 	return self:_playStopTrack("Stop", player_or_rig, path, false, fadeTime)
 end
 
 function AnimationsClass:GetTrackFromAlias(player_or_rig: Player | Model, alias: any): AnimationTrack?
 	self._initializedAssertion()
-	
+
 	return self:_getTrack(player_or_rig, alias, true)
 end
 
 function AnimationsClass:PlayTrackFromAlias(player_or_rig: Player | Model, alias: any, fadeTime: number?, weight: number?, speed: number?): AnimationTrack
 	self._initializedAssertion()
-	
+
 	return self:_playStopTrack("Play", player_or_rig, alias, true, fadeTime, weight, speed)
 end
 
 function AnimationsClass:StopTrackFromAlias(player_or_rig: Player | Model, alias: any, fadeTime: number?): AnimationTrack
 	self._initializedAssertion()
-	
+
 	return self:_playStopTrack("Stop", player_or_rig, alias, true, fadeTime)
 end
 
-function AnimationsClass:SetTrackAlias(player_or_rig: Player | Model, alias: any, path: any)
+function AnimationsClass:SetTrackAlias(player_or_rig: Player | Model, alias: any, path: {any} | string)
 	self._initializedAssertion()
-	
+
 	self:_setTrackAlias(player_or_rig, alias, path)
 end
 
 function AnimationsClass:RemoveTrackAlias(player_or_rig: Player | Model, alias: any)
 	self._initializedAssertion()
-	
+
 	self:_removeTrackAlias(player_or_rig, alias)
 end